Convert MZmine Data to mass_dataset Object

This function converts data from MZmine into a mass_dataset object. It processes the data to create a mass_dataset object containing expression data, sample information, and variable information.

Usage

convet_mzmine2mass_dataset(x, rt_unit = c("minute", "second"))

Arguments

x

A data frame containing MZmine data.

rt_unit

The unit for retention time, either "minute" or "second". Default is "minute".

Value

A mass_dataset object containing the processed MZmine data.

Details

The function reads MZmine data and processes it to create a mass_dataset object. It extracts sample information, variable information, and expression data. It also performs checks to ensure the data is correctly formatted.

Author

Xiaotao Shen xiaotao.shen@outlook.com

Examples

data("mzmine_table")
head(mzmine_table[,1:3])
#>   row ID   row m/z row retention time
#> 1      1 108.51790           10.24505
#> 2      2 184.98560           10.24529
#> 3      3  99.51304           10.24432
#> 4      4 176.97199           10.45647
#> 5      6 186.95641           10.47076
#> 6      7 108.51790           12.79046
object =
  convet_mzmine2mass_dataset(x = mzmine_table)
object
#> -------------------- 
#> massdataset version: 0.99.3 
#> -------------------- 
#> 1.expression_data:[ 3443 x 22 data.frame]
#> 2.sample_info:[ 22 x 2 data.frame]
#> 22 samples:10232_P4_RE4_01_476.mzXML Peak area 10765_P4_RE9_01_482.mzXML Peak area 10546_P4_RG11_01_515.mzXML Peak area ... 31878_P1_RH3_01_131.mzXML Peak area 14563_P1_RE10_01_99.mzXML Peak area
#> 3.variable_info:[ 3443 x 3 data.frame]
#> 3443 variables:1 2 3 ... 12248 12382
#> 4.sample_info_note:[ 2 x 2 data.frame]
#> 5.variable_info_note:[ 3 x 2 data.frame]
#> 6.ms2_data:[ 0 variables x 0 MS2 spectra]
#> -------------------- 
#> Processing information
#> 1 processings in total
#> create_mass_dataset ---------- 
#>       Package                Function.used                Time
#> 1 massdataset convet_mzmine2mass_dataset() 2026-03-04 12:54:19