Objects exported from other packages

These objects are imported from other packages. Follow the links below to see their documentation.

dplyr

arrange, arrange, case_when, count, count, count, desc, desc, desc, filter, filter, filter, glimpse, group_by, group_by, left_join, mutate, mutate, mutate_all, mutate_at, n, n, n, pull, relocate, relocate, rename, rename, rename_with, select, select, slice, slice, slice_head, slice_max, slice_min, slice_sample, slice_tail, summarise, summarize, summarize, transmute

ggplot2

aes, ggplot

magrittr

%>%

tidyr

drop_na, drop_na, fill, pivot_longer

Value

NULL No return value. This page is only for documentation purposes.