merge samples

merge samples

Usage

summarise_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

summarize_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

# S3 method for class 'mass_dataset'
summarise_samples(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  group_by,
  ...
)

Arguments

object

A mass_dataset object.

what

Summary statistic used to combine samples within each group.

group_by

Character vector giving one or more columns in sample_info used to define the grouping structure.

...

Additional arguments passed to downstream methods.

Value

A mass_dataset object with grouped samples collapsed into summary profiles and updated sample_info.

Examples

data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

object2 <- summarise_samples(
  object,
  what = "mean_intensity",
  group_by = "class"
)
head(extract_expression_data(object2))
#>                 Blank        QC Subject
#> M136T55_2_POS      NA 1447844.3 1988986
#> M79T35_POS         NA 2063212.8 2975154
#> M307T548_POS       NA  342037.7      NA
#> M183T224_POS       NA        NA      NA
#> M349T47_POS        NA 6417851.6 6975868
#> M182T828_POS  3167243        NA 5072689
head(extract_sample_info(object2))
#>   sample_id injection.order   class   group
#> 1     Blank               1   Blank   Blank
#> 2        QC               3      QC      QC
#> 3   Subject               5 Subject Subject